# perf: Increase data import's download template perf over 10x