This is about improving performance of a data import's download template functionality.

Here is the time taken to download 6000 user records through website's `download template`
```
06:48:49 web.1            | PATH: '/api/method/frappe.core.doctype.data_import.data_import.download_template'
06:48:49 web.1            |          145131537 function calls (145110247 primitive calls) in 108.133 seconds
```

With the new changes time taken is
```
07:46:27 web.1            | PATH: '/api/method/frappe.core.doctype.data_import.data_import.download_template'
07:46:27 web.1            |          6753111 function calls (6729791 primitive calls) in 7.565 seconds
```

Function calls reduced from 145million to 6million and time is reduced from 108 seconds to 7.5 secs.

**WHAT WAS THE ISSUE**
merge_data function @ https://github.com/frappe/frappe/blob/develop/frappe/core/doctype/data_import/exporter.py#L209 is taking almost all of the processing time. This function got 3 nested for loops and with the data size time of execution is also getting increased.

**HOW IS THAT FIXED**
Have removed nested for loops :)
Using generators where ever possible (This is to reduce memory usage)

**Issue**: ISS-20-21-10236